Commented/Added Screenshots/Uploaded README
